+++ src/wai/common/iterate/__init__.py
@@ -3,3 +3,4 @@ Utility functions for working with iterables and iterators.
 """
 from ._ConstantIterator import ConstantIterator
 from ._functions import safe_iter, flatten, invert_indices, extract_by_index, is_iterable, first, random
+from ._metadata import with_metadata, zip_metadata, get_metadata, unzip_metadata