+++ CHANGES.rst
@@ -28,3 +28,4 @@ Changelog
 - Added meta package, with functions to set and retrieve arbitrary meta-data against
   objects.
 - Added ConstantIterator class, which returns the same value over and over again.
+- Added metadata module to iterate, for working with metadata in iterables of objects.